Updated to GLFW: Removed SDL, FLyable camera implemented, modifications to shaders and Time

+++ src/ArchitectEngine/Time.cpp
@@ -1,3 +1,3 @@
 #include "Time.h"
 
-float Time::deltaTime;
+double Time::deltaTime;
